complete approval and reject request

+++ database/migrations/2022_06_09_164245_create_commodity_warehouse_table.php
@@ -17,7 +17,7 @@ class CreateCommodityWarehouseTable extends Migration
             $table->foreignId('commodity_id')->constrained();
             $table->foreignId('warehouse_id')->constrained();
             $table->primary(['commodity_id','warehouse_id']);
-            $table->integer('commodity_amount');
+            $table->double('commodity_amount');
             $table->timestamps();
         });
     }
